ViewModelが通信およびパースの処理の実装を知らなくても良いようにした
